slashed-project-engine/versatile-graphics-engine/requirements.txt
PyOpenGL
Pillow
numpy